Overhauled Sign in VC
